SharedFoldersSvc: Added new function for server-side file content copying (SHFL_FN_COPY_FILE_PART) and a function for query host features and supported functions (SHFL_FN_QUERY_FEATURES). [build fix]  bugref:9172
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@77838 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/HostServices/SharedFolders/vbsf.cpp
@@ -1381,7 +1381,7 @@ int vbsfCopyFilePart(SHFLCLIENTDATA *pClient, SHFLROOT idRootSrc, SHFLHANDLE hFi
 
     AssertPtrReturn(pClient, VERR_INVALID_PARAMETER);
 
-    size_t cbTotal = 0;
+    uint64_t cbTotal = 0;
 
     SHFLFILEHANDLE *pHandleSrc = vbsfQueryFileHandle(pClient, hFileSrc);
     int rc = vbsfCheckHandleAccess(pClient, idRootSrc, pHandleSrc, VBSF_CHECK_ACCESS_READ);
